Integrate TrackLibrary with MusicService to reduce complexity and to remove duplicate instance of TorrentSessionManager
